Move tests to the protocol package
